# src/services/report/routes/v1/public/doc/openapi.json
{
  "openapi": "3.0.0",
...
  "info": {
    "title": "ezReeport API",
    "version": "1.0",
    "contact": {
      "name": "ezTeam",
      "url": "https://github.com/ezpaarse-project",
      "email": "ezpaarse@couperin.org"
    },
    "license": {
      "name": "CeCILL",
      "url": "https://github.com/ezpaarse-project/ezreeport/blob/master/LICENSE.txt"
    },
    "description": "Reporting service"
  },
  "servers": [
    {
      "url": "/v1",
      "description": "Direct"
    },
    {
      "url": "/report/api/v1",
      "description": "ezMESURE"
    }
  ],
  "paths": {
...
    "/queues": {
      "get": {
        "summary": "Get all queues",
        "tags": [
          "queue"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "x-examples": {
                    "example-1": {
                      "status": {
                        "code": 200,
                        "message": "OK"
                      },
                      "content": [
                        "generation"
                      ]
                    }
                  },
                  "properties": {
                    "status": {
                      "type": "object",
                      "properties": {
                        "code": {
                          "type": "integer"
                        },
                        "message": {
                          "type": "string"
                        }
                      }
                    },
                    "content": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Queue"
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/400"
          },
          "401": {
            "$ref": "#/components/responses/401"
          },
          "403": {
            "$ref": "#/components/responses/403"
          },
          "500": {
            "$ref": "#/components/responses/500"
          }
        },
        "operationId": "queues-get",
        "description": "List all possible queues names.\n\nNeeds `general.queues-get` (see `/me/permissions` for more info).",
        "security": [
          {
            "User Token": []
          }
        ]
      }
    },
    "/queues/{queue}/pause": {
      "parameters": [
        {
          "schema": {
            "type": "string"
          },
          "name": "queue",
          "in": "path",
          "required": true,
          "description": "Queue name"
        }
      ],
      "put": {
        "summary": "Pause queue",
        "operationId": "queues-put-queue-pause",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "x-examples": {
                    "example-1": {
                      "status": {
                        "code": 200,
                        "message": "OK"
                      },
                      "content": {
                        "status": "active",
                        "jobs": []
                      }
                    }
                  },
                  "properties": {
                    "status": {
                      "type": "object",
                      "properties": {
                        "code": {
                          "type": "integer"
                        },
                        "message": {
                          "type": "string"
                        }
                      }
                    },
                    "content": {
                      "$ref": "#/components/schemas/Queue"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/400"
          },
          "401": {
            "$ref": "#/components/responses/401"
          },
          "403": {
            "$ref": "#/components/responses/403"
          },
          "404": {
            "$ref": "#/components/responses/404"
          },
          "500": {
            "$ref": "#/components/responses/500"
          }
        },
        "security": [
          {
            "User Token": []
          }
        ],
        "description": "Pause the whole queue.\n\nNeeds `general.queues-put-queue-pause` (see `/me/permissions` for more info).",
        "tags": [
          "queue"
        ]
      }
    },
    "/queues/{queue}/resume": {
      "parameters": [
        {
          "schema": {
            "type": "string"
          },
          "name": "queue",
          "in": "path",
          "required": true,
          "description": "Queue name"
        }
      ],
      "put": {
        "summary": "Resume queue",
        "operationId": "queues-put-queue-resume",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "x-examples": {
                    "example-1": {
                      "status": {
                        "code": 200,
                        "message": "OK"
                      },
                      "content": {
                        "status": "active",
                        "jobs": []
                      }
                    }
                  },
                  "properties": {
                    "status": {
                      "type": "object",
                      "properties": {
                        "code": {
                          "type": "integer"
                        },
                        "message": {
                          "type": "string"
                        }
                      }
                    },
                    "content": {
                      "$ref": "#/components/schemas/Queue"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/400"
          },
          "401": {
            "$ref": "#/components/responses/401"
          },
          "403": {
            "$ref": "#/components/responses/403"
          },
          "404": {
            "$ref": "#/components/responses/404"
          },
          "500": {
            "$ref": "#/components/responses/500"
          }
        },
        "description": "Resume the whole queue.\n\nNeeds `general.queues-put-queue-resume` (see `/me/permissions` for more info).",
        "security": [
          {
            "User Token": []
          }
        ],
        "tags": [
          "queue"
        ]
      }
    },
    "/queues/{queue}/jobs": {
      "parameters": [
        {
          "schema": {
            "type": "string"
          },
          "name": "queue",
          "in": "path",
          "required": true,
          "description": "Queue name"
        }
      ],
      "get": {
        "summary": "Get jobs of queue",
        "tags": [
          "queue"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "x-examples": {
                    "example-1": {
                      "status": {
                        "code": 200,
                        "message": "OK"
                      },
                      "content": {
                        "id": "135",
                        "data": {},
                        "progress": 1,
                        "added": "2022-11-15T11:01:33.179Z",
                        "started": "2022-11-15T11:01:33.184Z",
                        "ended": "2022-11-15T11:01:38.320Z",
                        "attempts": 1,
                        "status": "completed"
                      }
                    }
                  },
                  "properties": {
                    "status": {
                      "type": "object",
                      "properties": {
                        "code": {
                          "type": "integer"
                        },
                        "message": {
                          "type": "string"
                        }
                      }
                    },
                    "content": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Job"
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/401"
          },
          "401": {
            "$ref": "#/components/responses/401"
          },
          "403": {
            "$ref": "#/components/responses/403"
          },
          "404": {
            "$ref": "#/components/responses/404"
          },
          "500": {
            "$ref": "#/components/responses/500"
          }
        },
        "operationId": "queues-get-queue-jobId",
        "description": "Get jobs of the queue.\n\nNeeds `general.queues-get-queue-jobs` (see `/me/permissions` for more info).",
        "security": [
          {
            "User Token": []
          }
        ]
      }
    },
    "/queues/{queue}/jobs/{jobId}": {
      "parameters": [
        {
          "schema": {
            "type": "string"
          },
          "name": "queue",
          "in": "path",
          "required": true,
          "description": "Queue name"
        },
        {
          "schema": {
            "type": "string"
          },
          "name": "jobId",
          "in": "path",
          "required": true,
          "description": "Job id"
        }
      ],
      "get": {
        "summary": "Get specific job",
        "tags": [
          "queue"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "x-examples": {
                    "example-1": {
                      "status": {
                        "code": 200,
                        "message": "OK"
                      },
                      "content": {
                        "id": "135",
                        "data": {},
                        "progress": 1,
                        "added": "2022-11-15T11:01:33.179Z",
                        "started": "2022-11-15T11:01:33.184Z",
                        "ended": "2022-11-15T11:01:38.320Z",
                        "attempts": 1,
                        "status": "completed"
                      }
                    }
                  },
                  "properties": {
                    "status": {
                      "type": "object",
                      "properties": {
                        "code": {
                          "type": "integer"
                        },
                        "message": {
                          "type": "string"
                        }
                      }
                    },
                    "content": {
                      "$ref": "#/components/schemas/Job"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/401"
          },
          "401": {
            "$ref": "#/components/responses/401"
          },
          "403": {
            "$ref": "#/components/responses/403"
          },
          "404": {
            "$ref": "#/components/responses/404"
          },
          "500": {
            "$ref": "#/components/responses/500"
          }
        },
        "operationId": "queues-get-queue-jobs-jobId",
        "description": "Get specific job of the authed user's institution.\n\nNeeds `namespaces[namespaceId].queues-get` (see `/me/permissions` for more info).",
        "security": [
          {
            "User Token": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Task": {
        "type": "object",
        "x-examples": {},
        "description": "Reporting task",
        "x-tags": [
          "task"
        ],
        "additionalProperties": false,
        "properties": {
          "id": {
            "type": "string",
            "readOnly": true
          },
          "name": {
            "type": "string"
          },
          "namespaceId": {
            "type": "string",
            "description": "Namespace concerned by the task",
            "readOnly": true
          },
          "recurrence": {
            "type": "string",
            "enum": [
              "DAILY",
              "WEEKLY",
              "MONTHLY",
              "QUARTERLY",
              "BIENNIAL",
              "YEARLY"
            ],
            "description": "Defines when the task will be run."
          },
          "nextRun": {
            "type": "string",
            "description": "Next time the task will be run. Calculated if omiited",
            "format": "date-time"
          },
          "lastRun": {
            "type": "string",
            "format": "date-time",
            "description": "Last time the task has run",
            "nullable": true,
            "readOnly": true
          },
          "enabled": {
            "type": "boolean",
            "description": "Is the generation of the task is enabled",
            "default": true
          },
          "createdAt": {
            "type": "string",
            "readOnly": true
          },
          "updatedAt": {
            "type": "string",
            "nullable": true,
            "readOnly": true
          }
        },
        "required": [
          "name",
          "recurrence"
        ]
      },
...
      "Job": {
        "type": "object",
        "x-examples": {
          "example-1": {
            "id": "0",
            "data": {
              "...": "..."
            },
            "progress": 1,
            "added": "2022-11-15T11:01:33.179Z",
            "started": "2022-11-15T11:01:33.184Z",
            "ended": "2022-11-15T11:01:38.320Z",
            "attempts": 1,
            "status": "completed"
          }
        },
        "description": "Bull's Job",
        "properties": {
          "id": {
            "type": "string"
          },
          "data": {
            "type": "object",
            "description": "Data passed with Job",
            "properties": {
              "task": {
                "$ref": "#/components/schemas/Task"
              },
              "origin": {
                "type": "string"
              },
              "writeHistory": {
                "type": "boolean"
              },
              "debug": {
                "type": "boolean"
              }
            }
          },
          "result": {
            "type": "object"
          },
          "progress": {
            "type": "integer"
          },
          "added": {
            "type": "string"
          },
          "started": {
            "type": "string"
          },
          "ended": {
            "type": "string"
          },
          "attempts": {
            "type": "integer"
          },
          "status": {
            "type": "string"
          }
        }
      },
...
      "Queue": {
        "title": "Queue",
        "x-stoplight": {
          "id": "ukq73kbfu4p6d"
        },
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": [
              "active",
              "paused"
            ]
          }
        },
        "description": "Bull's Queue"
      },
...
    },
    "securitySchemes": {
      "User Token": {
        "type": "http",
        "scheme": "bearer",
        "description": "Used by user to interact with service"
      },
      "API Key": {
        "name": "X-API-Key",
        "type": "apiKey",
        "in": "header",
        "description": "Used by linked application to manage service"
      }
    },
...
    "responses": {
      "400": {
        "description": "Bad Request",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "status": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "integer"
                    },
                    "message": {
                      "type": "string"
                    }
                  }
                },
                "content": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    }
                  }
                }
              },
              "x-examples": {
                "example-1": {
                  "status": {
                    "code": 400,
                    "message": "Bad Request"
                  },
                  "content": {
                    "message": "JWT malformed: invalid signature"
                  }
                }
              }
            },
            "examples": {
              "Example": {
                "value": {
                  "status": {
                    "code": 400,
                    "message": "Bad Request"
                  },
                  "content": {
                    "message": "..."
                  }
                }
              }
            }
          }
        }
      },
      "401": {
        "description": "Unauthorized",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "status": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "integer"
                    },
                    "message": {
                      "type": "string"
                    }
                  }
                },
                "content": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    }
                  }
                }
              },
              "x-examples": {
                "example-1": {
                  "status": {
                    "code": 401,
                    "message": "Unauthorized"
                  },
                  "content": {
                    "message": "'GET /users/' requires auth"
                  }
                }
              }
            },
            "examples": {
              "Example": {
                "value": {
                  "status": {
                    "code": 401,
                    "message": "Unauthorized"
                  },
                  "content": {
                    "message": "'...' requires auth"
                  }
                }
              }
            }
          }
        }
      },
      "403": {
        "description": "Forbidden",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "status": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "integer"
                    },
                    "message": {
                      "type": "string"
                    }
                  }
                },
                "content": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    }
                  }
                }
              },
              "x-examples": {
                "example-1": {
                  "status": {
                    "code": 403,
                    "message": "Forbidden"
                  },
                  "content": {
                    "message": "User 'user' doesn't have the rights to acces to 'GET /users/'"
                  }
                }
              }
            },
            "examples": {
              "Example": {
                "value": {
                  "status": {
                    "code": 403,
                    "message": "Forbidden"
                  },
                  "content": {
                    "message": "User '...' doesn't have the rights to access to '...'"
                  }
                }
              }
            }
          }
        }
      },
      "404": {
        "description": "Not found",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "status": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "integer"
                    },
                    "message": {
                      "type": "string"
                    }
                  }
                },
                "content": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    }
                  }
                }
              },
              "x-examples": {
                "example-1": {
                  "status": {
                    "code": 404,
                    "message": "Not Found"
                  },
                  "content": {
                    "message": "User 'testa' not found"
                  }
                }
              }
            },
            "examples": {
              "Example": {
                "value": {
                  "status": {
                    "code": 404,
                    "message": "Not Found"
                  },
                  "content": {
                    "message": "... '...' not found"
                  }
                }
              }
            }
          }
        }
      },
...
      "500": {
        "description": "Internal Server Error",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "status": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "integer"
                    },
                    "message": {
                      "type": "string"
                    }
                  }
                },
                "content": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    }
                  }
                }
              },
              "x-examples": {
                "example-1": {
                  "status": {
                    "code": 500,
                    "message": "Internal Server Error"
                  },
                  "content": {
                    "message": "Unexpected token v in JSON at position 0"
                  }
                }
              }
            },
            "examples": {
              "Example": {
                "value": {
                  "status": {
                    "code": 500,
                    "message": "Internal Server Error"
                  },
                  "content": {
                    "message": "..."
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
...
}